chore(general): Deduplicate and move StatusWindow to StatusQ
