+++ ui/StatusQ/include/StatusQ/statuswindow.h
@@ -1,5 +1,4 @@
-#ifndef STATUSWINDOW_H
-#define STATUSWINDOW_H
+#pragma once
 
 #include <QQuickWindow>
 #include <QScreen>
@@ -11,31 +10,19 @@ class StatusWindow: public QQuickWindow
     Q_PROPERTY(bool isFullScreen READ isFullScreen NOTIFY isFullScreenChanged)
 
 public:
-
     explicit StatusWindow(QWindow *parent = nullptr);
 
     Q_INVOKABLE void toggleFullScreen();
+    Q_INVOKABLE void updatePosition();
 
     bool isFullScreen() const;
 
-    Q_INVOKABLE void updatePosition() {
-        auto point = QPoint(screen()->geometry().center().x() - geometry().width() / 2, screen()->geometry().center().y() - geometry().height() / 2);
-        if (point != this->position()) {
-            this->setPosition(point);
-        }
-    }
-
 signals:
     void isFullScreenChanged();
-    void secondInstanceDetected();
 
 private:
     void removeTitleBar();
     void showTitleBar();
-    void initCallbacks();
 
-private:
     bool m_isFullScreen;
 };
-
-#endif // STATUSWINDOW_H
