+++ ui/StatusQ/include/StatusQ/typesregistration.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void registerStatusQTypes();
