+++ ui/StatusQ/src/typesregistration.cpp
@@ -1,0 +1,10 @@
+#include "StatusQ/typesregistration.h"
+
+#include "StatusQ/statuswindow.h"
+
+#include <QQmlEngine>
+
+void registerStatusQTypes()
+{
+	qmlRegisterType<StatusWindow>("StatusQ", 0 , 1, "StatusWindow");
+}
